requirements.txt
fastapi
uvicorn
torch
torchvision
Pillow
numpy
boto3
